pyproject.toml
[project]
name = "living-document"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.11.0",
    "isort>=7.0.0",
    "langchain>=1.1.0",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.0",
    "langgraph>=1.0.4",
    "pydantic>=2.12.5",
    "pygithub>=2.8.1",
]
